requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pgvector==0.4.1
pydantic-settings==2.10.1
httpx==0.28.1
python-dotenv==1.1.1
pypdf==5.9.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-pptx==1.0.2
pytest==8.4.1
